detalles para proveedor

+++ InventarioyStock/src/main/java/com/IntegradorIO/InventarioyStock/Proveedor/ProveedorRepository.java
@@ -2,5 +2,6 @@ package com.IntegradorIO.InventarioyStock.Proveedor;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
+
 }
